bug fix: change int in frequency calculate in reduce method to float

+++ src/main/java/hk/ust/csit5970/BigramFrequencyPairs.java
@@ -98,7 +98,7 @@ public class BigramFrequencyPairs extends Configured implements Tool {
                 totalCount.set(sum);
                 VALUE.set(sum);
             } else {
-                VALUE.set(sum / totalCount.get());
+                VALUE.set((float) sum / (float) totalCount.get());
             }
             context.write(key, VALUE);
         }
